+++ doc/about.md
@@ -1,0 +1,1 @@
+# gotodo: Your CLI Task-Tracking Manager
