+++ main.go
@@ -1,10 +1,9 @@
 /*
-Copyright © 2023 NAME HERE <EMAIL ADDRESS>
-
+Copyright © 2023 dxZhang <dxzhang49@gmail.com>
 */
 package main
 
-import "github.com/danxuZhang/gotodo/cmd"
+import "github.com/danxuZhang/gotodo/src/cmd"
 
 func main() {
 	cmd.Execute()
