+++ src/cmd/joke.go
