+++ src/cmd/root.go
@@ -1,12 +1,11 @@
-/*
-Copyright © 2023 NAME HERE <EMAIL ADDRESS>
-*/
+// Package cmd
+// Copyright © 2023 dxZhang <dxzhang49@gmail.com>
+///*
 package cmd
 
 import (
-	"os"
-
 	"github.com/spf13/cobra"
+	"os"
 )
 
 // rootCmd represents the base command when called without any subcommands
